refactor: Modularization
