+++ .gitignore
@@ -2,7 +2,7 @@
 .idea
 .vscode
 *leak*
-indexer
+# indexer
 _scripts
 .postgres.sh
 wasm
