+++ mintbase-deps/src/common/payouts/mod.rs
@@ -1,0 +1,18 @@
+pub mod payout;
+pub mod royalty;
+pub mod splits;
+
+pub use payout::{
+    OwnershipFractions,
+    Payout,
+};
+pub use royalty::{
+    Royalty,
+    RoyaltyArgs,
+};
+pub use splits::{
+    NewSplitOwner,
+    SplitBetween,
+    SplitBetweenUnparsed,
+    SplitOwners,
+};
