+++ mintbase-deps/src/common/token.rs
@@ -1,0 +1,144 @@
+use std::collections::HashMap;
+
+use near_sdk::borsh::{
+    self,
+    BorshDeserialize,
+    BorshSerialize,
+};
+use near_sdk::serde::{
+    Deserialize,
+    Serialize,
+};
+use near_sdk::AccountId;
+
+use crate::common::{
+    ComposeableStats,
+    Loan,
+    Owner,
+    Royalty,
+    SplitOwners,
+    TokenKey,
+    TokenMetadataCompliant,
+};
+
+/// Supports NEP-171, 177, 178, 181. Ref:
+/// https://github.com/near/NEPs/blob/master/specs/Standards/NonFungibleToken/Core.md
+#[derive(Clone)]
+#[cfg_attr(feature = "wasm", derive(BorshDeserialize, BorshSerialize))]
+#[derive(Deserialize, Serialize)]
+pub struct Token {
+    /// The id of this token on this `Store`. Not unique across `Store`s.
+    /// `token_id`s count up from 0. Ref: https://github.com/near/NEPs/discussions/171
+    pub id: u64,
+    /// The current owner of this token. Either an account_id or a token_id (if composed).
+    pub owner_id: Owner,
+    /// Ref:
+    /// https://github.com/near/NEPs/blob/master/specs/Standards/NonFungibleToken/ApprovalManagement.md
+    /// Set of accounts that may transfer this token, other than the owner.
+    pub approvals: HashMap<AccountId, u64>,
+    /// The metadata content for this token is stored in the Contract
+    /// `token_metadata` field, to avoid duplication of metadata across tokens.
+    /// Use metadata_id to lookup the metadata. `Metadata`s is permanently set
+    /// when the token is minted.
+    pub metadata_id: u64,
+    /// The Royalty for this token is stored in the Contract `token_royalty`
+    /// field, to avoid duplication across tokens. Use royalty_id to lookup the
+    /// royalty. `Royalty`s are permanently set when the token is minted.
+    pub royalty_id: Option<u64>,
+    /// Feature for owner of this token to split the token ownership accross
+    /// several accounts.
+    pub split_owners: Option<SplitOwners>,
+    /// The account that minted this token.
+    pub minter: AccountId,
+    /// Non-nil if Token is loaned out. While token is loaned, disallow
+    /// transfers, approvals, revokes, etc. for the token, except from the
+    /// approved loan contract. Mark this field with the address of the loan
+    /// contract. See neps::loan for more.
+    pub loan: Option<Loan>,
+    /// Composeablility metrics for this token
+    pub composeable_stats: ComposeableStats,
+    /// If the token originated on another contract and was `nft_move`d to
+    /// this contract, this field will be non-nil.
+    pub origin_key: Option<TokenKey>,
+}
+
+impl Token {
+    /// - `metadata` validation performed in `TokenMetadataArgs::new`
+    /// - `royalty` validation performed in `Royalty::new`
+    pub fn new(
+        owner_id: AccountId,
+        token_id: u64,
+        metadata_id: u64,
+        royalty_id: Option<u64>,
+        split_owners: Option<SplitOwners>,
+        minter: AccountId,
+    ) -> Self {
+        Self {
+            owner_id: Owner::Account(owner_id),
+            id: token_id,
+            metadata_id,
+            royalty_id,
+            split_owners,
+            approvals: HashMap::new(),
+            minter,
+            loan: None,
+            composeable_stats: ComposeableStats::new(),
+            origin_key: None,
+        }
+    }
+
+    /// If the token is loaned, return the loaner as the owner.
+    pub fn get_owner_or_loaner(&self) -> Owner {
+        self.loan
+            .as_ref()
+            .map(|l| Owner::Account(l.holder.clone()))
+            .unwrap_or_else(|| self.owner_id.clone())
+    }
+
+    pub fn is_pred_owner(&self) -> bool {
+        self.owner_id.to_string() == near_sdk::env::predecessor_account_id().to_string()
+    }
+
+    pub fn is_loaned(&self) -> bool {
+        self.loan.is_some()
+    }
+}
+
+// Supports NEP-171, 177, 178, 181. Ref:
+/// https://github.com/near/NEPs/blob/master/specs/Standards/NonFungibleToken/Core.md
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub struct TokenCompliant {
+    /// The id of this token on this `Store`. Not unique across `Store`s.
+    /// `token_id`s count up from 0. Ref: https://github.com/near/NEPs/discussions/171
+    pub id: u64,
+    /// The current owner of this token. Either an account_id or a token_id (if composed).
+    pub owner_id: Owner,
+    /// Ref:
+    /// https://github.com/near/NEPs/blob/master/specs/Standards/NonFungibleToken/ApprovalManagement.md
+    /// Set of accounts that may transfer this token, other than the owner.
+    pub approvals: HashMap<AccountId, u64>,
+    /// The metadata content for this token is stored in the Contract
+    /// `token_metadata` field, to avoid duplication of metadata across tokens.
+    /// Use metadata_id to lookup the metadata. `Metadata`s is permanently set
+    /// when the token is minted.
+    pub metadata: TokenMetadataCompliant,
+    /// The Royalty for this token is stored in the Contract `token_royalty`
+    /// field, to avoid duplication across tokens. Use royalty_id to lookup the
+    /// royalty. `Royalty`s are permanently set when the token is minted.
+    pub royalty: Option<Royalty>,
+    /// Feature for owner of this token to split the token ownership accross
+    /// several accounts.
+    pub split_owners: Option<SplitOwners>,
+    /// The account that minted this token.
+    pub minter: AccountId,
+    /// Non-nil if Token is loaned out. While token is loaned, disallow
+    /// transfers, approvals, revokes, etc. for the token, except from the
+    /// approved loan contract. Mark this field with the address of the loan
+    /// contract. See neps::loan for more.
+    pub loan: Option<Loan>,
+    /// Composeablility metrics for this token
+    pub composeable_stats: ComposeableStats,
+    /// If the token originated on another contract and was `nft_move`d to
+    /// this contract, this field will be non-nil.
+    pub origin_key: Option<TokenKey>,
+}
