+++ mintbase-deps/src/lib.rs
@@ -1,16 +1,11 @@
-mod consts;
-mod enums;
-mod fns;
-mod impls;
-mod structs;
-mod traits;
+pub mod common;
+pub mod consts;
+pub mod interfaces;
+pub mod logging;
 
-pub use consts::*;
-pub use enums::*;
-pub use fns::*;
-pub use impls::*;
-pub use structs::*;
-pub use traits::*;
+// re-export for market
+#[cfg(feature = "wasm")]
+pub use near_sdk;
 
 #[cfg(feature = "factory-wasm")]
 pub mod factory;
@@ -19,60 +14,12 @@ pub mod helper;
 #[cfg(feature = "store-wasm")]
 pub mod store;
 
-pub mod logging;
+#[cfg(feature = "all")]
+pub mod indexer;
+// TODO: end global scope pollution
+#[cfg(feature = "all")]
+pub use consts::*;
+#[cfg(feature = "all")]
+pub use indexer::*;
 
-#[cfg(feature = "all")]
-mod mintbase_std {
-    pub use {
-        bigdecimal,
-        chrono,
-        dotenv,
-        futures,
-        hyper,
-        near_account_id,
-        near_client_primitives,
-        near_crypto,
-        near_indexer,
-        near_jsonrpc_client,
-        near_jsonrpc_primitives,
-        near_primitives,
-        near_sdk,
-        openssl_probe,
-        rayon,
-        reqwest,
-        tokio,
-        tokio_postgres,
-        tokio_stream,
-        tower,
-        tracing,
-        tracing_appender,
-        tracing_subscriber,
-        uuid,
-    };
-}
-#[cfg(feature = "all")]
-pub use logging::{
-    Nep171EventLog,
-    NftApproveLog,
-    NftBurnLog,
-    NftComposeLog,
-    NftListLog,
-    NftLoanSetLog,
-    NftMarketLog,
-    NftMintLog,
-    NftMintLogMemo,
-    NftOfferLog2,
-    NftOptionStringLog,
-    NftRevokeLog,
-    NftSaleLog,
-    NftSetSplitOwnerLog,
-    NftStoreCreateLog,
-    NftStringLog,
-    NftTransferLog,
-    NftUpdateListLog,
-    NftUpdateOfferLog,
-};
-#[cfg(feature = "all")]
-pub use mintbase_std::*;
-#[cfg(feature = "wasm")]
-pub use near_sdk;
+pub mod utils;
